pyproject.toml
[tool.poetry]
name = "db-connector"
version = "0.1.0"
description = ""
authors = ["Bastiancito"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
psycopg2-binary = "^2.9.10"
pyodbc = "^5.2.0"
numpy = "^2.2.1"
sqlalchemy = "^2.0.36"
pandas = "^2.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
